Upgrade specs2 to 3.6.6.
Yup. Specs2 3. Weird, huh? Anyway, I'm hoping it'll fix our test error
on Jenkins.

+++ test/controllers/api/DocumentSetControllerSpec.scala
@@ -47,7 +47,7 @@ class DocumentSetControllerSpec extends ApiControllerSpecification {
 
     "return a DocumentSet" in new CreateScope {
       val json = contentAsString(result)
-      json must /("documentSet") /("id" -> documentSet.id)
+      json must /("documentSet") /("id" -> documentSet.id.toInt)
       json must /("documentSet") /("title" -> documentSet.title)
     }
